refactor: move `findLastAssignmentAction()` into common
